% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osc_read.R
\name{tosc_read}
\alias{tosc_read}
\title{Read a touchOSC layout file}
\usage{
tosc_read(tosc)
}
\arguments{
\item{tosc}{a file with the extension .touchosc}
}
\value{
an xml2 object
}
\description{
Read a touchOSC layout file
}
\examples{
basic <- system.file("extdata", "basic.touchosc", package = "toscr")
x <- tosc_read(basic)
xml2::xml_find_all(x, ".//control")
}
